Add namespace `Textual`.
