+++ docs/textual/A.md
@@ -93,8 +93,8 @@ $anchor->template('<span>{tag}</span>');
 
 ## Attributes
 
-Refer to the [Attribute Tests](https://github.com/php-forge/html/blob/main/tests/A/AttributeTest.php) for comprehensive
-examples.
+Refer to the [Attribute Tests](https://github.com/php-forge/html/blob/main/tests/Textual/A/AttributeTest.php) for
+comprehensive examples.
 
 The following methods are available for setting attributes:
 
@@ -134,7 +134,7 @@ The following methods are available for setting attributes:
 
 ## Custom methods
 
-Refer to the [Custom Method Test](https://github.com/php-forge/html/blob/main/tests/A/CustomMethodTest.php) for
+Refer to the [Custom Method Test](https://github.com/php-forge/html/blob/main/tests/Textual/A/CustomMethodTest.php) for
 comprehensive examples.
 
 The following methods are available for customizing the `HTML` output:
