+++ docs/textual/Span.md
@@ -1,14 +1,19 @@
-# I
+# Span
 
-The `<i>` `HTML` element represents a range of text that's set off from the normal text for some reason, such as an
-idiomatic text, technical terms, taxonomic designations, among others.
+The `<span>` `HTML` element represents a generic inline container for phrasing content, which doesn't inherently
+represent anything. It can be used to group elements for styling purposes (using the class or id attributes), or
+because they share attribute values, such as lang. It should be used only when no other semantic element is
+appropriate.
+ 
+`<span>` is very much like a `<div>` element, but `<div>` is a block-level element whereas a `<span>` is an inline 
+element.
 
 ## Basic Usage
 
-Instantiate the `I` class using `I::widget()`.
+Instantiate the `Span` class using `Span::widget()`.
 
 ```php
-$i = I::widget();
+$span = Span::widget();
 ```
 
 ## Setting Attributes
@@ -17,35 +22,35 @@ Use the provided methods to set specific attributes for the a element.
 
 ```php
 // setting class attribute
-$i->class('fa fa-home');
+$span->class('container');
 ```
 
 Or, use the `attributes` method to set multiple attributes at once.
 
 ```php
-$i->attributes(['class' => 'fa fa-home', 'aria-hidden' => 'true']);
+$span->attributes(['class' => 'container', 'style' => 'background-color: #eee;']);
 ```
 
 ## Adding Content
 
-If you want to include content within the `i` tag, use the `content` method.
+If you want to include content within the `div` tag, use the `content` method.
 
 ```php
-$i->content('->');
+$span->content('MyContent');
 ```
 
 ## Rendering
 
-Generate the `HTML` output using the `render` method.
+Generate the `HTML` output using the `render` method, for simple instantiation. 
 
 ```php
-$html = $i->render();
+$html = $span->render();
 ```
 
 Or, use the magic `__toString` method.
 
 ```php
-$html = (string) $i;
+$html = (string) $span;
 ```
 
 ## Common Use Cases
@@ -54,47 +59,18 @@ Below are examples of common use cases:
 
 ```php
 // adding multiple attributes
-$i->class('fa fa-home')->title('Home');
+$span->class('external')->content('MyContent');
 
 // using data attributes
-$i->dataAttributes(['bs-toggle' => 'modal', 'bs-target' => '#exampleModal', 'analytics' => 'trackClick']);
+$span->dataAttributes(['analytics' => 'trackClick']);
 ```
 
-Explore additional methods for setting various attributes such as `lang`, `tabindex`, `title`, and more.
-
-## Prefix and Suffix
-
-Use `prefix` and `suffix` methods to add text before and after the `i` tag, respectively.
-
-```php
-// adding a prefix
-$html = $i->content('->')->prefix('Next')->render();
-
-// adding a suffix
-$html = $i->content('<-')->suffix('Previous')->render();
-```
-
-## Template
-
-The `template` method allows you to customize the `HTML` output of the a element.
-
-The following template tags are available:
-
-| Tag        | Description      |
-| ---------- | ---------------- |
-| `{prefix}` | The prefix text. |
-| `{tag}`    | The a element.   |
-| `{suffix}` | The suffix text. |
-
-```php
-// using a custom template
-$i->template('<span>{tag}</span>');
-```
+Explore additional methods for setting various attributes such as `lang`, `name`, `style`, `title`, etc.
 
 ## Attributes
 
-Refer to the [Attribute Tests](https://github.com/php-forge/html/blob/main/tests/I/AttributeTest.php) for comprehensive
-examples.
+Refer to the [Attribute Tests](https://github.com/php-forge/html/blob/main/tests/Textual/Span/AttributeTest.php) for
+comprehensive examples.
 
 The following methods are available for setting attributes:
 
@@ -102,7 +78,7 @@ The following methods are available for setting attributes:
 | ----------------- | ------------------------------------------------------------------------------------------------ |
 | `attributes()`    | Set multiple `attributes` at once.                                                               |
 | `class()`         | Set the `class` attribute.                                                                       |
-| `content()`       | Set the `content` within the `i` element.                                                        |
+| `content()`       | Set the `content` within the `div` element.                                                      |
 | `dataAttributes()`| Set multiple `data-attributes` at once.                                                          |
 | `id()`            | Set the `id` attribute.                                                                          |
 | `lang()`          | Set the `lang` attribute.                                                                        |
@@ -112,23 +88,23 @@ The following methods are available for setting attributes:
 
 ## Custom methods
 
-Refer to the [Custom Method Test](https://github.com/php-forge/html/blob/main/tests/I/CustomMethodTest.php) for 
-comprehensive examples.
+Refer to the [Custom Methods Tests](https://github.com/php-forge/html/blob/main/tests/Textual/Span/CustomMethodTest.php)
+for comprehensive examples.
 
 The following methods are available for customizing the `HTML` output:
 
 | Method                       | Description                                                                           |
 | ---------------------------- | ------------------------------------------------------------------------------------- |
-| `prefix()`                   | Add text before the `i` element.                                                      |
+| `prefix()`                   | Add text before the `a` element.                                                      |
 | `prefixContainer()`          | Set enabled or disabled for the `prefix-container` element.                           |
 | `prefixContainerAttributes()`| Set `attributes` for the `prefix-container` element.                                  |                                            
 | `prefixContainerClass()`     | Set the `class` attribute for the `prefix-container` element.                         |
 | `prefixContainerTag()`       | Set the `tag` for the `prefix-container` element.                                     |
 | `render()`                   | Generates the `HTML` output.                                                          |
-| `suffix()`                   | Add text after the `i` element.                                                       |
+| `suffix()`                   | Add text after the `a` element.                                                       |
 | `suffixContainer()`          | Set enabled or disabled for the `suffix-container` element.                           |
 | `suffixContainerAttributes()`| Set `attributes` for the `suffix-container` element.                                  |
 | `suffixContainerClass()`     | Set the `class` attribute for the `suffix-container` element.                         |
 | `suffixContainerTag()`       | Set the `tag` for the `suffix-container` element.                                     |
-| `template()`                 | Set the `template` for the `i` element.                                               |
-| `widget()`                   | Instantiates the `I::class`.                                                          |
+| `template()`                 | Set the template for the `HTML` output.                                               |
+| `widget()`                   | Instantiates the `Span::class`.                                                       |
