+++ tests/FormControl/Form/AttributeTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace PHPForge\Html\Tests\FormControl\Form;
 
 use PHPForge\Html\FormControl\Form;
-use PHPForge\Html\Span;
+use PHPForge\Html\Textual\Span;
 use PHPForge\Support\Assert;
 use PHPUnit\Framework\TestCase;
